requirements.txt
streamlit>=1.33
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow>=10.0.0
